requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
yfinance>=0.1.63
ta-lib>=0.4.24 
python-dotenv>=0.19.0

tensorflow>=2.8.0
keras>=2.8.0

aiohttp>=3.8.1
asyncio>=3.4.3

sqlite3 

talib-binary>=0.4.19 

tqdm>=4.62.0
colorama>=0.4.4
matplotlib>=3.4.3
seaborn>=0.11.2

logging  

argparse  

pytz>=2021.3
python-dateutil>=2.8.2

requests>=2.26.0

joblib>=1.1.0
schedule>=1.1.0

ujson>=4.0.2
numba>=0.54.1